creating boilerplate API for dashboard and stock creation
