+++ Insomnia_2020-08-14.json
@@ -1,0 +1,1 @@
+{"_type":"export","__export_format":4,"__export_date":"2020-08-14T09:44:49.622Z","__export_source":"insomnia.desktop.app:v2020.2.0","resources":[{"_id":"req_3208792e30e740218d99db29a248b1f6","authentication":{},"body":{"mimeType":"application/json","text":"{\n\t\"stock\": \"APPL\"\n}"},"created":1597398178217,"description":"","headers":[{"id":"pair_59d9aadf02c3449e80741a2e3e114b06","name":"Content-Type","value":"application/json"}],"isPrivate":false,"metaSortKey":-1597398178217,"method":"POST","modified":1597398273502,"name":"Creating Stock","parameters":[],"parentId":"fld_50771126cb9b4becbbadb0c11c8c4e91","settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingFollowRedirects":"global","settingRebuildPath":true,"settingSendCookies":true,"settingStoreCookies":true,"url":"localhost:8000/stock","_type":"request"},{"_id":"fld_50771126cb9b4becbbadb0c11c8c4e91","created":1597398078661,"description":"","environment":{},"environmentPropertyOrder":null,"metaSortKey":-1597398078661,"modified":1597398078661,"name":"Stock Screener","parentId":"wrk_693354065b9e404c8f005bb236160519","_type":"request_group"},{"_id":"wrk_693354065b9e404c8f005bb236160519","created":1587309317414,"description":"","modified":1587309317414,"name":"Insomnia","parentId":null,"scope":null,"_type":"workspace"},{"_id":"env_ab9a3f44a21de93e0323a238003141d5041ea4b2","color":null,"created":1587309317486,"data":{},"dataPropertyOrder":null,"isPrivate":false,"metaSortKey":1587309317486,"modified":1587309317486,"name":"Base Environment","parentId":"wrk_693354065b9e404c8f005bb236160519","_type":"environment"},{"_id":"jar_ab9a3f44a21de93e0323a238003141d5041ea4b2","cookies":[{"creation":"2020-04-19T15:16:44.063Z","domain":"localhost","expires":"2021-04-23T15:24:21.000Z","extensions":["SameSite=Lax"],"hostOnly":true,"id":"43219492581984964","key":"csrftoken","lastAccessed":"2020-04-24T15:24:21.186Z","maxAge":31449600,"path":"/","value":"PEKiDTO47vOUnoxyBnBGrMzvkGBMOb3RrLjP2AGJ1g6dI1e4bm3iK6u8MivxanbS"}],"created":1587309317490,"modified":1587741861186,"name":"Default Jar","parentId":"wrk_693354065b9e404c8f005bb236160519","_type":"cookie_jar"},{"_id":"spc_be6b6754d49348f4b9ecf1a45340d6cc","contentType":"yaml","contents":"","created":1597398019541,"fileName":"Insomnia","modified":1597398019541,"parentId":"wrk_693354065b9e404c8f005bb236160519","_type":"api_spec"},{"_id":"env_27ab004f239c4d4c8f86525ef2179480","color":null,"created":1597398106850,"data":{},"dataPropertyOrder":null,"isPrivate":false,"metaSortKey":1597398106850,"modified":1597398106850,"name":"New Environment","parentId":"env_ab9a3f44a21de93e0323a238003141d5041ea4b2","_type":"environment"}]}
