+++ Makefile
@@ -1,2 +1,5 @@
+make helloworld:
+	uvicorn hello_world:app --reload
+
 make server:
 	uvicorn main:app --reload
