+++ pyproject.toml
@@ -10,6 +10,9 @@ python = "^3.8"
 black = "^19.10b0"
 fastapi = "^0.61.0"
 uvicorn = "^0.11.8"
+sqlalchemy = "1.3.17"
+jinja2 = "^2.11.2"
+yfinance = "^0.1.54"
 
 [tool.poetry.dev-dependencies]
